ASoC: jack: integration with extcon for snd-jack
This patch integration with extcon for snd-jack. We previously
used snd-jack only with input driver. But, we can also use this
snd-jack with extcon driver.
However, this doesn't mean we must use extcon driver unconditionally
to use snd-jack. There isn't a dependency between snd-jack and extcon
in this patch. if CONFIG_EXTCON is not defined, extcon API will just
return properly, So, It isn't influence on snd-jack.

Change-Id: Ifa856ad62a31301effdfbe90d88757c0cca42c36
Signed-off-by: Inha Song <ideal.song@samsung.com>