ASoC: max98090: Change to enough time for microphone detection
This patch change to enough time for microphone detection in Maxim98090
codec driver.

To detect microphone in max98090 codec, JDWK bit of Jack detect configuration
register should be set to 0.

In codec driver, Set the JDWK bit to 0 for microphone detection and, after
some delay, set the JDWK bit to 1 to reduce power consumption.
To detect micrphone certainly, we need enough time to delay.

Change-Id: I33c444b580b4efe6b6d5d72ad093eb548d1ccfb8
Signed-off-by: Inha Song <ideal.song@samsung.com>